Add previews to Asciinema players
Times are taken from Yarik's PR (#26).

I intentionally chose to /not/ add a preview for the MRI Data Management page.
It is ~6 minutes and the preview takes ~5 seconds to generate (utilizing 100%
CPU). The other pages don't seem to be affected by this, so I'm assuming it's
due to length (or the complexity of the cast).